+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Current
 
+  - Added new WYSIWYG editor [Jodit](https://xdsoft.net/jodit/) as a replacement for SCEditor.
+  - Ace Editor, SimpleMDE and SCEditor can now be configured using global config options which can be overridden for single properties using schema options.   
+  - Refactored src/string.js and moved hardcoded support for special editors (Ace Editor, SimpleMDE and SCEditor) into seperate editor files which then extends src/string.js.
   - Added [Choices.js](https://github.com/jshjohnson/Choices) support for array and select editors.
   - Added 2 new config options: "disable_select2" and "disable_selectize" which can be used to disable select2/selectize on a specific property/field.
   - Removed "rating" editor and updated "starrating" editor to accept/emulate "rating" editor options.